Stratified estimator for Ms.GSI

Stratified estimator for Ms.GSI

Usage

stratified_estimator_msgsi(
  mdl_out = NULL,
  path = NULL,
  mixvec,
  catchvec,
  cv = NULL
)

Arguments

mdl_out

Optional. Model output object name.

path

Where to find output from each mixture as a folder.

mixvec

Character vector of mixture sillies that are used to locate the folders where output .csv files lives, if mdl_out is not provided.

catchvec

Numeric vector of harvest for each mixture/stratum, must be in the same order as mixvec.

cv

Numeric vector of harvest estimate coefficients of variation for each stratum, must be the same order as mixvec.

Value

A tibble of proportions and harvest numbers by reporting group for combined mixtures/strata.

Examples

stratified_estimator_msgsi(path = "test", mixvec = c("m1", "m2", "m3"), catchvec = c(4500, 5000, 3000), cv = c(0.5, 0.1 ,0.3))
#> Error in stratified_estimator_msgsi(path = "test", mixvec = c("m1", "m2",     "m3"), catchvec = c(4500, 5000, 3000), cv = c(0.5, 0.1, 0.3)): `Either provide a Ms.GSI output or a valid `path` to folders containing output .csv files for each mixture.